Set a stack limit of 9500
D does not have StackOverflowExceptions like Java but instead just gives a
segmentation fault, which is not very informative. eval.d now contains a new
class for that. StackOverflowExceptions are not caught in every function and
just stop Lyra outright after printing the internal callstack.
The callstack does not include special commands like 'if, 'define, etc.